requirements.txt
# Backend
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
pydantic[email]==2.7.1

# ML Engine
numpy==1.26.4
scikit-learn==1.5.0

# PostgreSQL (uncomment for production)
# psycopg2-binary==2.9.9
